+++ Sources/Sentry/SentryHttpTransport.m
@@ -1,6 +1,4 @@
 #import "SentryHttpTransport.h"
-#import "SentryDataCategory.h"
-#import "SentryDataCategoryMapper.h"
 #import "SentryDiscardReasonMapper.h"
 #import "SentryEnvelopeItemHeader.h"
 #import "SentryEnvelopeRateLimit.h"
@@ -85,8 +83,9 @@
         typeof(self) __weak weakSelf = self;
         [self.fileManager
             setEnvelopeDeletedCallback:^(SentryEnvelopeItem *item, NSUInteger category) {
-                [weakSelf envelopeItemDeleted:item
-                                 withCategory:sentryDataCategoryForNSUInteger(category)];
+                [weakSelf
+                    envelopeItemDeleted:item
+                           withCategory:[SentryDataCategoryMapper categoryForNSUInteger:category]];
             }];
 
         [self sendAllCachedEnvelopes];
@@ -143,24 +142,21 @@
     [self.fileManager storeEnvelope:envelope];
 }
 
-- (void)recordLostEvent:(NSUInteger)category reason:(NSUInteger)reason
+- (void)recordLostEvent:(SentryDataCategory)category reason:(SentryDiscardReason)reason
 {
     [self recordLostEvent:category reason:reason quantity:1];
 }
 
-- (void)recordLostEvent:(NSUInteger)categoryRaw
-                 reason:(NSUInteger)reasonRaw
+- (void)recordLostEvent:(SentryDataCategory)category
+                 reason:(SentryDiscardReason)reason
                quantity:(NSUInteger)quantity
 {
     if (!self.sendClientReports) {
         return;
     }
 
-    SentryDataCategory category = (SentryDataCategory)categoryRaw;
-    SentryDiscardReason reason = (SentryDiscardReason)reasonRaw;
-
-    NSString *key = [NSString stringWithFormat:@"%@:%@", nameForSentryDataCategory(category),
-        nameForSentryDiscardReason(reason)];
+    NSString *key = [NSString stringWithFormat:@"%@:%@",
+        [SentryDataCategoryMapper nameFor:category], nameForSentryDiscardReason(reason)];
 
     @synchronized(self.discardedEvents) {
         SentryDiscardedEvent *event = self.discardedEvents[key];
@@ -168,9 +164,10 @@
             quantity = event.quantity + quantity;
         }
 
-        event = [[SentryDiscardedEvent alloc] initWithReason:nameForSentryDiscardReason(reason)
-                                                    category:nameForSentryDataCategory(category)
-                                                    quantity:quantity];
+        event =
+            [[SentryDiscardedEvent alloc] initWithReason:nameForSentryDiscardReason(reason)
+                                                category:[SentryDataCategoryMapper nameFor:category]
+                                                quantity:quantity];
 
         self.discardedEvents[key] = event;
     }
@@ -194,7 +191,7 @@
     @synchronized(self) {
         if (_isFlushing) {
             SENTRY_LOG_DEBUG(@"Already flushing.");
-            return SentryFlushResultAlreadyFlushing;
+            return kSentryFlushResultAlreadyFlushing;
         }
 
         SENTRY_LOG_DEBUG(@"Start flushing.");
@@ -225,10 +222,10 @@
 
     if (result == 0) {
         SENTRY_LOG_DEBUG(@"Finished flushing.");
-        return SentryFlushResultSuccess;
+        return kSentryFlushResultSuccess;
     } else {
         SENTRY_LOG_WARN(@"Flushing timed out.");
-        return SentryFlushResultTimedOut;
+        return kSentryFlushResultTimedOut;
     }
 }
 
@@ -250,18 +247,18 @@
                withCategory:(SentryDataCategory)dataCategory;
 {
     SENTRY_LOG_WARN(@"Envelope item dropped due to exceeding rate limit. Category: %@",
-        nameForSentryDataCategory(dataCategory));
-    [self recordLostEvent:dataCategory reason:kSentryDiscardReasonRateLimitBackoff];
-    [self recordLostSpans:envelopeItem reason:kSentryDiscardReasonRateLimitBackoff];
-    [self recordLostLogBytes:envelopeItem reason:kSentryDiscardReasonRateLimitBackoff];
+        [SentryDataCategoryMapper nameFor:dataCategory]);
+    [self recordLostEvent:dataCategory reason:SentryDiscardReasonRateLimitBackoff];
+    [self recordLostSpans:envelopeItem reason:SentryDiscardReasonRateLimitBackoff];
+    [self recordLostLogBytes:envelopeItem reason:SentryDiscardReasonRateLimitBackoff];
 }
 
 - (void)envelopeItemDeleted:(SentryEnvelopeItem *)envelopeItem
                withCategory:(SentryDataCategory)dataCategory
 {
-    [self recordLostEvent:dataCategory reason:kSentryDiscardReasonCacheOverflow];
-    [self recordLostSpans:envelopeItem reason:kSentryDiscardReasonCacheOverflow];
-    [self recordLostLogBytes:envelopeItem reason:kSentryDiscardReasonCacheOverflow];
+    [self recordLostEvent:dataCategory reason:SentryDiscardReasonCacheOverflow];
+    [self recordLostSpans:envelopeItem reason:SentryDiscardReasonCacheOverflow];
+    [self recordLostLogBytes:envelopeItem reason:SentryDiscardReasonCacheOverflow];
 }
 
 #pragma mark private methods
@@ -471,10 +468,11 @@
         if ([itemType isEqualToString:SentryEnvelopeItemTypes.clientReport]) {
             continue;
         }
-        SentryDataCategory category = sentryDataCategoryForEnvelopItemType(itemType);
-        [self recordLostEvent:category reason:kSentryDiscardReasonSendError];
-        [self recordLostSpans:item reason:kSentryDiscardReasonSendError];
-        [self recordLostLogBytes:item reason:kSentryDiscardReasonSendError];
+        SentryDataCategory category =
+            [SentryDataCategoryMapper categoryForEnvelopeItemType:itemType];
+        [self recordLostEvent:category reason:SentryDiscardReasonSendError];
+        [self recordLostSpans:item reason:SentryDiscardReasonSendError];
+        [self recordLostLogBytes:item reason:SentryDiscardReasonSendError];
     }
 }
 
@@ -482,7 +480,7 @@
 {
     if ([SentryEnvelopeItemTypes.log isEqualToString:envelopeItem.type]) {
         NSUInteger byteCount = envelopeItem.data.length;
-        [self recordLostEvent:kSentryDataCategoryLogByte reason:reason quantity:byteCount];
+        [self recordLostEvent:SentryDataCategoryLogByte reason:reason quantity:byteCount];
     }
 }
 
@@ -498,7 +496,7 @@
             return;
         }
         NSArray *spans = transactionJson[@"spans"] ?: [NSArray array];
-        [self recordLostEvent:kSentryDataCategorySpan reason:reason quantity:spans.count + 1];
+        [self recordLostEvent:SentryDataCategorySpan reason:reason quantity:spans.count + 1];
     }
 }