+++ Sources/SentryObjCCompat/SentryObjCDataCollectionHttpBodyType.swift
@@ -1,3 +1,4 @@
+#if SDK_V10
 // swiftlint:disable missing_docs
 #if SWIFT_PACKAGE
 internal import SentrySwift
@@ -18,3 +19,4 @@ extension SentryObjCDataCollectionHttpBodyType {
 }
 
 // swiftlint:enable missing_docs
+#endif // SDK_V10