+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Breaking Changes
+
+- Only expose `experimental.dataCollection` APIs in SDK V10 (#8435)
+
 ## 9.22.0
 
 ### Features
@@ -7,6 +13,7 @@
 - Mark opt in option `swiftAsyncStacktraces` as stable (#8373)
 - Add `dataCollection` option under `experimental` for configuring data scrubbing behavior (#8369)
 - Add scope API to clear feature flags (#8364)
+- Add dictionary initialization for `options.experimental.dataCollection` (#8371)
 
 ### Fixes